Standardized "flags" fields as u8, simplified conversions and using the right reference where appropriate

+++ src/expressions/operations.rs
@@ -1,35 +1,35 @@
 use crate::types::enums::AsmErrorEquates;
 
 // value, flags, lastWasOp
-pub type ResultValues = (i64, i32, bool);
+pub type ResultValues = (i64, u8, bool);
 pub type ResultOfOperation = Result<ResultValues, AsmErrorEquates>;
-pub type ExpressionOperationFunc = fn(v1: i64, v2: i64, f1: i32, f2: i32) -> ResultOfOperation;
+pub type ExpressionOperationFunc = fn(v1: i64, v2: i64, f1: u8, f2: u8) -> ResultOfOperation;
 
-pub fn take_least_significant_byte(v1: i64, _v2: i64, f1: i32, _f2: i32) -> ResultOfOperation {
+pub fn take_least_significant_byte(v1: i64, _v2: i64, f1: u8, _f2: u8) -> ResultOfOperation {
 	Ok((v1 & 0xff as i64, f1, false))
 }
 
-pub fn take_most_significant_byte(v1: i64, _v2: i64, f1: i32, _f2: i32)-> ResultOfOperation {
+pub fn take_most_significant_byte(v1: i64, _v2: i64, f1: u8, _f2: u8)-> ResultOfOperation {
 	Ok((v1 >> 8 & 0xff, f1, false))
 }
 
-pub fn negate(v1: i64, _v2: i64, f1: i32, _f2: i32)-> ResultOfOperation {
+pub fn negate(v1: i64, _v2: i64, f1: u8, _f2: u8)-> ResultOfOperation {
 	Ok((-v1, f1, false))
 }
 
-pub fn invert(v1: i64, _v2: i64, f1: i32, _f2: i32)-> ResultOfOperation {
+pub fn invert(v1: i64, _v2: i64, f1: u8, _f2: u8)-> ResultOfOperation {
 	Ok((!v1, f1, false))
 }
 
-pub fn not(v1: i64, _v2: i64, f1: i32, _f2: i32)-> ResultOfOperation {
+pub fn not(v1: i64, _v2: i64, f1: u8, _f2: u8)-> ResultOfOperation {
 	Ok(((v1 == 0) as i64, f1, false))
 }
 
-pub fn multiply(v1: i64, v2: i64, f1: i32, f2: i32) -> ResultOfOperation {
+pub fn multiply(v1: i64, v2: i64, f1: u8, f2: u8) -> ResultOfOperation {
 	Ok((v1 * v2, f1 | f2, true))
 }
 
-pub fn divide(v1: i64, v2: i64, f1: i32, f2: i32) -> ResultOfOperation {
+pub fn divide(v1: i64, v2: i64, f1: u8, f2: u8) -> ResultOfOperation {
 	// Conversion note: in the original C code, this sets lastWasOp *before* stackarg() executes,
 	// which means it's ignored, since it's reset to "false" inside stackarg(). I believe this is
 	// a bug in the original code (since multiply, adding, and subtracting *do* set it to "true")
@@ -44,7 +44,7 @@ pub fn divide(v1: i64, v2: i64, f1: i32, f2: i32) -> ResultOfOperation {
 	}
 }
 
-pub fn modulo(v1: i64, v2: i64, f1: i32, f2: i32) -> ResultOfOperation {
+pub fn modulo(v1: i64, v2: i64, f1: u8, f2: u8) -> ResultOfOperation {
 	if f1 | f2 != 0 {
 		// Conversion note: in the original, only this path doesn't enable lastWasOp, but I believe this is a bug.
 		Ok((0, f1 | f2, true))
@@ -55,7 +55,7 @@ pub fn modulo(v1: i64, v2: i64, f1: i32, f2: i32) -> ResultOfOperation {
 	}
 }
 
-pub fn question(v1: i64, v2: i64, f1: i32, f2: i32) -> ResultOfOperation {
+pub fn question(v1: i64, v2: i64, f1: u8, f2: u8) -> ResultOfOperation {
 	if f1 != 0 {
 		Ok((0, f1, false))
 	} else {
@@ -75,15 +75,15 @@ pub fn question(v1: i64, v2: i64, f1: i32, f2: i32) -> ResultOfOperation {
 	}
 }
 
-pub fn add(v1: i64, v2: i64, f1: i32, f2: i32) -> ResultOfOperation {
+pub fn add(v1: i64, v2: i64, f1: u8, f2: u8) -> ResultOfOperation {
 	Ok((v1 + v2, f1 | f2, true))
 }
 
-pub fn subtract(v1: i64, v2: i64, f1: i32, f2: i32) -> ResultOfOperation {
+pub fn subtract(v1: i64, v2: i64, f1: u8, f2: u8) -> ResultOfOperation {
 	Ok((v1 - v2, f1 | f2, true))
 }
 
-pub fn shift_right(v1: i64, v2: i64, f1: i32, f2: i32) -> ResultOfOperation {
+pub fn shift_right(v1: i64, v2: i64, f1: u8, f2: u8) -> ResultOfOperation {
 	if f1 | f2 != 0 {
 		Ok((0, f1 | f2, false))
 	} else {
@@ -91,7 +91,7 @@ pub fn shift_right(v1: i64, v2: i64, f1: i32, f2: i32) -> ResultOfOperation {
 	}
 }
 
-pub fn shift_left(v1: i64, v2: i64, f1: i32, f2: i32) -> ResultOfOperation {
+pub fn shift_left(v1: i64, v2: i64, f1: u8, f2: u8) -> ResultOfOperation {
 	if f1 | f2 != 0 {
 		Ok((0, f1 | f2, false))
 	} else {
@@ -99,31 +99,31 @@ pub fn shift_left(v1: i64, v2: i64, f1: i32, f2: i32) -> ResultOfOperation {
 	}
 }
 
-pub fn greater(v1: i64, v2: i64, f1: i32, f2: i32) -> ResultOfOperation {
+pub fn greater(v1: i64, v2: i64, f1: u8, f2: u8) -> ResultOfOperation {
 	Ok(((v1 > v2) as i32 as i64, f1 | f2, false))
 }
 
-pub fn greater_or_equal(v1: i64, v2: i64, f1: i32, f2: i32) -> ResultOfOperation {
+pub fn greater_or_equal(v1: i64, v2: i64, f1: u8, f2: u8) -> ResultOfOperation {
 	Ok(((v1 >= v2) as i32 as i64, f1 | f2, false))
 }
 
-pub fn lesser(v1: i64, v2: i64, f1: i32, f2: i32) -> ResultOfOperation {
+pub fn lesser(v1: i64, v2: i64, f1: u8, f2: u8) -> ResultOfOperation {
 	Ok(((v1 < v2) as i32 as i64, f1 | f2, false))
 }
 
-pub fn lesser_or_equal(v1: i64, v2: i64, f1: i32, f2: i32) -> ResultOfOperation {
+pub fn lesser_or_equal(v1: i64, v2: i64, f1: u8, f2: u8) -> ResultOfOperation {
 	Ok(((v1 <= v2) as i32 as i64, f1 | f2, false))
 }
 
-pub fn equal(v1: i64, v2: i64, f1: i32, f2: i32) -> ResultOfOperation {
+pub fn equal(v1: i64, v2: i64, f1: u8, f2: u8) -> ResultOfOperation {
 	Ok(((v1 == v2) as i32 as i64, f1 | f2, false))
 }
 
-pub fn not_equal(v1: i64, v2: i64, f1: i32, f2: i32) -> ResultOfOperation {
+pub fn not_equal(v1: i64, v2: i64, f1: u8, f2: u8) -> ResultOfOperation {
 	Ok(((v1 != v2) as i32 as i64, f1 | f2, false))
 }
 
-pub fn and_and(v1: i64, v2: i64, f1: i32, f2: i32) -> ResultOfOperation {
+pub fn and_and(v1: i64, v2: i64, f1: u8, f2: u8) -> ResultOfOperation {
 	if f1 == 0 && v1 == 0 || f2 == 0 && v2 == 0 {
 		Ok((0, 0, false))
 	} else {
@@ -131,7 +131,7 @@ pub fn and_and(v1: i64, v2: i64, f1: i32, f2: i32) -> ResultOfOperation {
 	}
 }
 
-pub fn or_or(v1: i64, v2: i64, f1: i32, f2: i32) -> ResultOfOperation {
+pub fn or_or(v1: i64, v2: i64, f1: u8, f2: u8) -> ResultOfOperation {
 	if f1 == 0 && v1 != 0 || f2 == 0 && v2 != 0 {
 		Ok((1, 0, false))
 	} else {
@@ -139,14 +139,14 @@ pub fn or_or(v1: i64, v2: i64, f1: i32, f2: i32) -> ResultOfOperation {
 	}
 }
 
-pub fn xor(v1: i64, v2: i64, f1: i32, f2: i32) -> ResultOfOperation {
+pub fn xor(v1: i64, v2: i64, f1: u8, f2: u8) -> ResultOfOperation {
 	Ok((v1 ^ v2, f1 | f2, false))
 }
 
-pub fn and(v1: i64, v2: i64, f1: i32, f2: i32) -> ResultOfOperation {
+pub fn and(v1: i64, v2: i64, f1: u8, f2: u8) -> ResultOfOperation {
 	Ok((v1 & v2, f1 | f2, false))
 }
 
-pub fn or(v1: i64, v2: i64, f1: i32, f2: i32) -> ResultOfOperation {
+pub fn or(v1: i64, v2: i64, f1: u8, f2: u8) -> ResultOfOperation {
 	Ok((v1 | v2, f1 | f2, false))
 }
